Don't add a trailing NULL to the XAUTH password.
Currently, racoon on Android adds a NULL byte at the end of the
XAUTH password. Various servers have workarounds for this, but
they also support the correct behaviour of not having the NULL
character.

isakmp_xauth_req strips the NULL byte from the end of
xauth->login, but does not strip it from the end of xauth->pass.
Strip it in set_xauth_and_more instead of in the ISAKMP xauth
code because set_xauth_and_more is only called directly from
setup() which parses argv, whereas the ISAKMP code can get the
password via other more complicated means.

Bug: 17251660
Change-Id: I5c1ff91d39862c8262b2eae7cd304ae52048f7cb